backend/requirements.txt
fastapi==0.117.1
uvicorn==0.36.0
python-multipart==0.0.20
pycryptodome==3.23.0
pydantic==2.11.9
